Disable incompatible test when running karma tests
