+++ test/node/test_full_bson.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var BSON = require('../..'),
-  fs = require('fs'),
   Buffer = require('buffer').Buffer,
   BinaryParser = require('../binary_parser').BinaryParser,
   ObjectID = BSON.ObjectID,
@@ -676,19 +675,6 @@ describe('Full BSON', function() {
     var serialized_data = bson.serialize(doc);
     var deserialized_data = bson.deserialize(serialized_data);
     expect(doc.doc.value()).to.equal(deserialized_data.doc.value());
-    done();
-  });
-
-  it('Should Correctly Deserialize bson file from mongodump', function(done) {
-    var data = fs.readFileSync('test/node/data/test.bson', { encoding: null });
-    var docs = [];
-    var bsonIndex = 0;
-    while (bsonIndex < data.length)
-      bsonIndex = bson.deserializeStream(data, bsonIndex, 1, docs, docs.length, {
-        isArray: true
-      });
-
-    expect(docs.length).to.equal(1);
     done();
   });
